Time: 0 ms (100.00%), Space: 6.1 MB (70.00%) - LeetHub

+++ 2177-find-three-consecutive-integers-that-sum-to-a-given-number/2177-find-three-consecutive-integers-that-sum-to-a-given-number.cpp
@@ -6,7 +6,7 @@ public:
         
         if((num % 3) == 0)
         {
-            long long val = (long long)(num / 3);
+            long long val =(num / 3);
             res.push_back(val-1);
             res.push_back(val);
             res.push_back(val+1);
